refactor Genre into /data
