+++ app/src/androidTest/java/com/github/fribourgsdp/radio/AddPlaylistActivityTest.kt
@@ -19,6 +19,7 @@ import androidx.test.espresso.matcher.ViewMatchers.isDisplayed
 import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import com.github.fribourgsdp.radio.data.Genre
 import com.github.fribourgsdp.radio.data.User
 import com.google.android.gms.tasks.Tasks
 import org.hamcrest.Matchers
