+++ app/src/main/java/com/github/fribourgsdp/radio/GameActivity.kt
@@ -1,7 +1,5 @@
 package com.github.fribourgsdp.radio
 
-import android.app.Activity
-import android.app.Dialog
 import android.content.Intent
 import android.os.Bundle
 import android.view.*
