+++ app/src/main/java/com/github/fribourgsdp/radio/MainActivity.kt
@@ -2,12 +2,9 @@ package com.github.fribourgsdp.radio
 
 
 
-import android.content.ContentValues
 import android.content.Intent
 import android.os.Bundle
-import android.util.Log
 import android.widget.Button
-import androidx.appcompat.app.AppCompatActivity
 
 //import com.github.fribourgsdp.radio.databinding.ActivityMainBinding
