+++ app/src/main/java/com/github/fribourgsdp/radio/data/User.kt
@@ -6,8 +6,6 @@ import com.github.fribourgsdp.radio.AppSpecificFileSystem
 import com.github.fribourgsdp.radio.FileSystem
 import com.github.fribourgsdp.radio.FirestoreDatabase
 import com.github.fribourgsdp.radio.SetUtility
-import com.github.fribourgsdp.radio.data.Playlist
-import com.github.fribourgsdp.radio.data.Song
 
 import com.google.android.gms.tasks.Task
 import com.google.android.gms.tasks.Tasks
